Blocking detection

+++ goPatch/src/advocatego/advocate_tracing.go
@@ -17,6 +17,8 @@ import (
 	"runtime"
 	"strconv"
 	"time"
+
+	_ "unsafe"
 )
 
 var traceFileCounter = 0
@@ -32,17 +34,22 @@ var startWriting = false
 
 // InitTracing initializes the tracing.
 // The function creates the trace folder and starts the background memory test.
-func InitTracing(timeout int) {
+//
+//go:linkname InitTracing runtime.AdvocateInitTracing
+func InitTracing(timeout int, init bool) {
+	if initRun { // called by main but alredy run by init
+		return
+	}
+	initRun = true
+
 	startTime = time.Now()
 	timerStarted = true
 
 	// remove the trace folder if it exists
 	err := os.RemoveAll(tracePathRecorded)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			println("Cannot remove: ", err.Error())
-			return
-		}
+	if err != nil && !os.IsNotExist(err) {
+		println("Cannot remove: ", err.Error())
+		return
 	}
 
 	// create the trace folder
@@ -66,12 +73,14 @@ func InitTracing(timeout int) {
 
 	// go writeTraceIfFull()
 	// go removeAtomicsIfFull()
-	runtime.InitTracing(FinishTracing, WriteToTraceFile)
+	runtime.InitTracing(FinishTracing, WriteToTraceFile, init)
 }
 
 // Write the trace of the program to a file.
 // The trace is written in the file named file_name.
 // The trace is written in the format of advocate.
+//
+//go:linkname FinishTracing runtime.AdvocateFinishTracing
 func FinishTracing() {
 	if hasFinished {
 		// needed to prevent program stop while still writing
